Rename duty_cycle to duty_ratio
There is quite a lot of confusion (at least in my mind) about the duty
cycle. Every reference I can find says that it is the high time vs
period, but in the 555 datasheets that I have found the ratio seems to
be tl / (tl + th), which gives the low time.

I have now renamed it to duty ratio to reflect the fact that it is the
high time ratio.